Move connectivity help to own module
