+++ common/auth/src/main/kotlin/app/ss/auth/AuthRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,12 +30,12 @@ import app.ss.auth.api.toModel
 import app.ss.models.auth.SSUser
 import app.ss.network.safeApiCall
 import app.ss.storage.db.dao.UserDao
-import com.cryart.sabbathschool.core.extensions.connectivity.ConnectivityHelper
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 import retrofit2.Response
+import ss.foundation.android.connectivity.ConnectivityHelper
 import ss.foundation.coroutines.DispatcherProvider
 import timber.log.Timber
 import javax.inject.Inject
