+++ common/auth/src/main/kotlin/app/ss/auth/api/TokenAuthenticator.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,13 +25,13 @@ package app.ss.auth.api
 import app.ss.network.NetworkResource
 import app.ss.network.safeApiCall
 import app.ss.storage.db.dao.UserDao
-import com.cryart.sabbathschool.core.extensions.connectivity.ConnectivityHelper
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 import okhttp3.Authenticator
 import okhttp3.Request
 import okhttp3.Response
 import okhttp3.Route
+import ss.foundation.android.connectivity.ConnectivityHelper
 import ss.foundation.coroutines.DispatcherProvider
 import javax.inject.Inject
 import javax.inject.Singleton
