+++ common/lessons-data/src/main/java/app/ss/lessons/data/repository/lessons/LessonInfoDataSource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -29,8 +29,8 @@ import app.ss.models.SSLesson
 import app.ss.models.SSLessonInfo
 import app.ss.storage.db.dao.LessonsDao
 import app.ss.storage.db.entity.LessonEntity
-import com.cryart.sabbathschool.core.extensions.connectivity.ConnectivityHelper
 import com.cryart.sabbathschool.core.response.Resource
+import ss.foundation.android.connectivity.ConnectivityHelper
 import ss.foundation.coroutines.DispatcherProvider
 import ss.lessons.api.SSLessonsApi
 import javax.inject.Inject
