+++ common/lessons-data/src/main/java/app/ss/lessons/data/repository/lessons/ReadsDataSource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,8 +28,8 @@ import app.ss.lessons.data.repository.LocalDataSource
 import app.ss.models.SSRead
 import app.ss.storage.db.dao.ReadsDao
 import app.ss.storage.db.entity.ReadEntity
-import com.cryart.sabbathschool.core.extensions.connectivity.ConnectivityHelper
 import com.cryart.sabbathschool.core.response.Resource
+import ss.foundation.android.connectivity.ConnectivityHelper
 import ss.foundation.coroutines.DispatcherProvider
 import ss.lessons.api.SSLessonsApi
 import javax.inject.Inject
