+++ common/lessons-data/src/test/java/app/ss/lessons/data/repository/quarterly/LanguagesDataSourceTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,12 +23,12 @@
 package app.ss.lessons.data.repository.quarterly
 
 import app.ss.storage.db.dao.LanguagesDao
-import com.cryart.sabbathschool.core.extensions.connectivity.ConnectivityHelper
 import io.mockk.every
 import io.mockk.mockk
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Before
 import org.junit.Test
+import ss.foundation.android.connectivity.ConnectivityHelper
 import ss.foundation.coroutines.test.TestDispatcherProvider
 import ss.lessons.api.SSQuarterliesApi
