+++ features/lessons/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -65,6 +65,7 @@ dependencies {
     implementation(projects.features.media)
     implementation(projects.features.pdf)
     implementation(projects.features.reader)
+    implementation(projects.libraries.foundation.android)
     implementation(projects.libraries.foundation.coroutines)
 
     implementation(libs.google.material)
