+++ libraries/foundation/android/src/main/kotlin/ss/foundation/android/connectivity/ConnectivityManager.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.cryart.sabbathschool.core.extensions.connectivity
+package ss.foundation.android.connectivity
 
 import android.content.Context
 import android.net.ConnectivityManager
