+++ libraries/lessons/api/build.gradle.kts
@@ -29,6 +29,7 @@ plugins {
 android { namespace = "ss.lessons.api" }
 
 dependencies {
+    api(projects.libraries.foundation.coroutines)
     api(projects.libraries.lessons.model)
     api(libs.square.okhttp)
     api(libs.square.retrofit)
