+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -69,6 +69,7 @@ include(
     ":features:settings",
     ":libraries:circuit-helpers:api",
     ":libraries:circuit-helpers:impl",
+    ":libraries:foundation:android",
     ":libraries:foundation:coroutines",
     ":libraries:foundation:coroutines:test",
     ":libraries:lessons:api",
